pyproject.toml
[project]
name = "academicservice"
version = "0.1.0"
description = "Academic chatbot support knowledgebase, chating and realtime calling"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "deepgram-sdk>=3.7.7",
    "langchain-community>=0.3.11",
    "langgraph>=0.2.59",
    "neo4j>=5.27.0",
    "openai>=1.57.2",
    "pymongo[srv]>=4.10.1",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=3.3.1",
    "starlette>=0.41.3",
    "uvicorn[standard]>=0.32.1",
    "tavily-python",
    "beautifulsoup4",
]

